+++ lib/features/usernames/presentation/usernames_screen.dart
@@ -3,7 +3,6 @@ import 'package:anonaddy/common/constants/app_strings.dart';
 import 'package:anonaddy/common/created_at_widget.dart';
 import 'package:anonaddy/common/custom_app_bar.dart';
 import 'package:anonaddy/common/error_message_widget.dart';
-import 'package:anonaddy/common/list_tiles/alias_detail_list_tile.dart';
 import 'package:anonaddy/common/list_tiles/recipient_list_tile.dart';
 import 'package:anonaddy/common/offline_banner.dart';
 import 'package:anonaddy/common/platform_aware_widgets/dialogs/platform_alert_dialog.dart';
@@ -15,6 +14,7 @@ import 'package:anonaddy/common/utilities.dart';
 import 'package:anonaddy/features/associated_aliases/presentation/associated_aliases.dart';
 import 'package:anonaddy/features/recipients/presentation/recipients_screen.dart';
 import 'package:anonaddy/features/usernames/domain/username.dart';
+import 'package:anonaddy/features/usernames/presentation/components/alias_detail_list_tile.dart';
 import 'package:anonaddy/features/usernames/presentation/controller/usernames_screen_notifier.dart';
 import 'package:anonaddy/features/usernames/presentation/controller/usernames_screen_state.dart';
 import 'package:anonaddy/features/usernames/presentation/username_default_recipient.dart';
